Move Job specific code to the blockchain/job.go
Move Job and Escrow contracts specific code to the blockchain/job.go and
blockchain/escrow.go respectively.
