+++ blockchain/blockchain.go
@@ -8,47 +8,16 @@ import (
 	"crypto/ecdsa"
 	"crypto/md5"
 	"encoding/hex"
-	"encoding/json"
 
 	"github.com/coreos/bbolt"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/pkg/errors"
 	"github.com/singnet/snet-daemon/config"
-	"github.com/singnet/snet-daemon/db"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/grpc"
-)
-
-const (
-	jobPendingState = "PENDING"
-	jobFundedState  = "FUNDED"
-
-	// PaymentTypeHeader is a type of payment used to pay for a RPC call.
-	// Supported types are: "job", "escrow".
-	PaymentTypeHeader = "snet-payment-type"
-	// JobPaymentType each call should be payed using unique instance of funded Job
-	JobPaymentType = "job"
-	// EscrowPaymentType each call should have id and nonce of payment channel
-	// in metadata.
-	EscrowPaymentType = "escrow"
-
-	JobAddressHeader   = "snet-job-address"
-	JobSignatureHeader = "snet-job-signature"
-
-	// PaymentChannelIdHeader is a MultiPartyEscrow contract payment channel id
-	PaymentChannelIdHeader = "snet-payment-channel-id"
-	// PaymentChannelNonceHeader is a payment channel nonce value
-	PaymentChannelNonceHeader = "snet-payment-channel-nonce"
-	// PaymentChannelAmountHeader is an amount of payment channel value
-	// which server is authorized withdraw after handling the RPC call.
-	PaymentChannelAmountHeader = "snet-payment-channel-amount"
-	// PaymentChannelSignatureHeader is a signature of the client to confirm
-	// authorized amount
-	PaymentChannelSignatureHeader = "snet-payment-channel-signature"
 )
 
 var (
@@ -150,95 +119,4 @@ func (p Processor) GrpcStreamInterceptor() grpc.StreamServerInterceptor {
 	}
 
 	return noOpInterceptor
-}
-
-func (p Processor) IsValidJobInvocation(jobAddressBytes, jobSignatureBytes []byte) bool {
-	log := log.WithFields(log.Fields{
-		"jobAddress":   common.BytesToAddress(jobAddressBytes).Hex(),
-		"jobSignature": hex.EncodeToString(jobSignatureBytes)})
-
-	// Get job state from db
-	log.Debug("retrieving job from database")
-	job := &db.Job{}
-
-	p.boltDB.View(func(tx *bolt.Tx) error {
-		bucket := tx.Bucket(db.JobBucketName)
-		jobBytes := bucket.Get(jobAddressBytes)
-		if jobBytes != nil {
-			json.Unmarshal(jobBytes, job)
-		}
-		return nil
-	})
-
-	// If job is marked completed locally, reject
-	if job.Completed {
-		log.Debug("job already marked completed locally")
-		return false
-	}
-
-	v, r, s, err := parseSignature(jobSignatureBytes)
-	if err != nil {
-		log.WithError(err).Error("error parsing signature")
-		return false
-	}
-
-	pubKey, err := crypto.SigToPub(p.sigHasher(jobAddressBytes), bytes.Join([][]byte{jobSignatureBytes[0:64], {v % 27}},
-		[]byte{}))
-	if err != nil {
-		log.WithError(err).Error("error recovering signature")
-		return false
-	}
-
-	// If job is FUNDED and signature validates, skip on-chain validation
-	if job.JobState == jobFundedState && bytes.Equal(crypto.PubkeyToAddress(*pubKey).Bytes(), job.Consumer) {
-		log.Debug("validated job invocation locally")
-		return true
-	}
-
-	log.Debug("unable to validate job invocation locally; falling back to on-chain validation")
-
-	// Fall back to on-chain validation
-	if validated, err := p.agent.ValidateJobInvocation(&bind.CallOpts{
-		Pending: true,
-		From:    common.HexToAddress(p.address)}, common.BytesToAddress(jobAddressBytes), v, r, s); err != nil {
-		log.WithError(err).Error("error validating job on chain")
-		return false
-	} else if !validated {
-		log.Debug("job failed to validate")
-		return false
-	}
-
-	log.Debug("validated job invocation on chain")
-	return true
-}
-
-func (p Processor) CompleteJob(jobAddressBytes, jobSignatureBytes []byte) {
-	job := &db.Job{}
-
-	// Mark the job completed in the db synchronously
-	if err := p.boltDB.Update(func(tx *bolt.Tx) error {
-		bucket := tx.Bucket(db.JobBucketName)
-		jobBytes := bucket.Get(jobAddressBytes)
-		if jobBytes != nil {
-			json.Unmarshal(jobBytes, job)
-		}
-		job.Completed = true
-		job.JobSignature = jobSignatureBytes
-		jobBytes, err := json.Marshal(job)
-		if err != nil {
-			return err
-		}
-		if err = bucket.Put(jobAddressBytes, jobBytes); err != nil {
-			return err
-		}
-		return nil
-	}); err != nil {
-		log.WithFields(log.Fields{
-			"jobAddress":   common.BytesToAddress(jobAddressBytes).Hex(),
-			"jobSignature": hex.EncodeToString(jobSignatureBytes),
-		}).WithError(err).Error("error marking job completed in db")
-	}
-
-	// Submit the job for completion
-	p.jobCompletionQueue <- &jobInfo{jobAddressBytes, jobSignatureBytes}
 }
